// x-pack/solutions/observability/plugins/exploratory_view/public/components/shared/exploratory_view/types.d.ts
export declare const ReportViewTypes: {
    readonly dist: "data-distribution";
    readonly kpi: "kpi-over-time";
    readonly cwv: "core-web-vitals";
    readonly mdd: "device-data-distribution";
    readonly smt: "single-metric";
    readonly htm: "heatmap";
};
type ValueOf<T> = T[keyof T];
export type ReportViewTypeId = keyof typeof ReportViewTypes;
export type ReportViewType = ValueOf<typeof ReportViewTypes>;
export interface ColumnFilter {
    language: 'kuery';
    query: string;
}
export interface ParamFilter {
    label: string;
    input: ColumnFilter;
}
export interface MetricOption {
    id: string;
    field?: string;
    label: string;
    description?: string;
    columnType?: 'range' | 'operation' | 'FILTER_RECORDS' | 'TERMS_COLUMN' | 'unique_count' | typeof FORMULA_COLUMN;
    columnFilters?: ColumnFilter[];
    columnFilter?: ColumnFilter;
    paramFilters?: ParamFilter[];
    timeScale?: string;
    showPercentileAnnotations?: boolean;
    formula?: string;
    metricStateOptions?: Pick<MetricState, 'colorMode' | 'palette' | 'titlePosition' | 'textAlign'>;
    palette?: PaletteOutput;
    format?: 'percent' | 'number';
    emptyAsNull?: boolean;
    timestampField?: string;
}
export interface SeriesConfig {
    reportType: ReportViewType | string;
    xAxisColumn: Partial<LastValueIndexPatternColumn> | Partial<DateHistogramIndexPatternColumn>;
    yAxisColumns: Array<Partial<FieldBasedIndexPatternColumn>>;
    breakdownFields: string[];
    defaultSeriesType: SeriesType;
    filterFields: Array<string | {
        field: string;
        nested?: string;
        isNegated?: boolean;
    }>;
    seriesTypes: SeriesType[];
    baseFilters?: Array<PersistableFilter | ExistsFilter | PhraseFilter>;
    definitionFields: Array<string | {
        field: string;
        nested?: string;
        singleSelection?: boolean;
        filters?: Array<PersistableFilter | ExistsFilter | PhraseFilter>;
    }>;
    textDefinitionFields?: string[];
    metricOptions?: Array<MetricOption | {
        id: string;
        field?: string;
        label: string;
        items: MetricOption[];
        columnType?: string;
    }>;
    labels: Record<string, string>;
    hasOperationType: boolean;
    palette?: PaletteOutput;
    yTitle?: string;
    yConfig?: YConfig[];
    query?: {
        query: string;
        language: 'kuery';
    };
}
export type URLReportDefinition = Record<string, string[]>;
export type URLTextReportDefinition = Record<string, string>;
export interface SeriesUrl {
    name: string;
    time: {
        to: string;
        from: string;
    };
    breakdown?: string;
    filters?: UrlFilter[];
    seriesType?: SeriesType;
    operationType?: OperationType;
    dataType: AppDataType;
    reportDefinitions?: URLReportDefinition;
    textReportDefinitions?: URLTextReportDefinition;
    selectedMetricField?: string;
    hidden?: boolean;
    showPercentileAnnotations?: boolean;
    color?: string;
}
export interface UrlFilter {
    field: string;
    values?: Array<string | number>;
    notValues?: Array<string | number>;
    wildcards?: string[];
    notWildcards?: string[];
}
export interface ConfigProps {
    dataView?: DataView;
    series?: SeriesUrl;
    spaceId?: string;
}
interface FormatType extends SerializedFieldFormat<FieldFormatParams> {
    id: 'duration' | 'number' | 'bytes' | 'percent';
}
export type AppDataType = 'synthetics' | 'uptime' | 'ux' | 'infra_logs' | 'infra_metrics' | 'apm' | 'mobile' | 'alerts';